+++ psr_lines.c
@@ -4,11 +4,10 @@
 #include <string.h>
 #include <math.h>
 #include "psrgeom.h"
+#include <time.h>
 
-// Define arbitrary numbers for grid types
-#define  GT_CART  1
-#define  GT_CYL2  2
-#define  GT_CYL3  3
+// Generate a random number between -1 and 1
+#define RANDU  ((double)rand()/(double)RAND_MAX * 2.0 - 1.0)
 
 struct opts
 {
@@ -17,6 +16,7 @@ struct opts
     double  P_sec;     // period, in sec
     double  tmult;     // RK4 step size, as a fraction of lt cyl radius
     int     rL_norm;   // bool: normalise to light cylinder radius?
+    double  rho_lim;    // limit to within rho_lim * light cylinder radius?
     char   *outfile;   // name of output file (NULL means stdout)
 };
 
@@ -26,6 +26,9 @@ void print_col_headers( FILE *f );
 
 int main( int argc, char *argv[] )
 {
+    // Seed random number generator
+    srand( time( NULL ) );
+
     // Set up struct for command line options and set default values
     struct opts o;
     o.al_deg    = NAN;
@@ -33,6 +36,7 @@ int main( int argc, char *argv[] )
     o.ze_deg    = NAN;
     o.tmult     = 0.01;
     o.rL_norm   = 0;
+    o.rho_lim   = 1.2;
     o.outfile   = NULL;
 
     parse_cmd_line( argc, argv, &o );
@@ -66,23 +70,27 @@ int main( int argc, char *argv[] )
 
     set_pulsar( &psr, ra, dec, P, r, al, ze );
 
-    // Set up grid point
-    point X;
-    point far_pt;
+    // Set up points
+    point X;  // The starting point
+    point X2; // The end point of either footpoint() or farpoint()
 
     print_psrg_header( f, argc, argv );
 
     // Set the point at this gridpoint
-    set_point_xyz( &X, 0.4*psr.rL,
-                       0.4*psr.rL,
-                       0.4*psr.rL,
+    set_point_xyz( &X, RANDU*psr.rL,
+                       RANDU*psr.rL,
+                       RANDU*psr.rL,
                        POINT_SET_ALL );
 
     // Write the column headers
     print_col_headers( f );
 
     // Calculate the points along the magnetic field line
-    farpoint( &X, &psr, o.tmult, f, o.rL_norm, &far_pt );
+    footpoint( &X, &psr, o.tmult, DIR_INWARD, f, o.rL_norm, o.rho_lim, &X2 );
+    fprintf( f, "\n\n" );
+    footpoint( &X, &psr, o.tmult, DIR_OUTWARD, f, o.rL_norm, o.rho_lim, &X2 );
+    fprintf( f, "\n\n" );
+    farpoint( &X, &psr, o.tmult, f, o.rL_norm, o.rho_lim, &X2 );
 
     // Clean up
     destroy_psr_angle( ra  );
@@ -113,6 +121,8 @@ void usage()
     printf( "  -L           Normalise distances to light cylinder radius\n" );
     printf( "  -o  outfile  The name of the output file to write to. If not "
                            "set, output will be written to stdout.\n" );
+    printf( "  -r  rho_lim  The maximum distance allowed for lines. "
+                           "(default = 1.2)\n" );
     printf( "  -t  tmult    The initial size of the RK4 steps, as a fraction "
                            "of the light cylinder radius (default: 0.01)\n" );
 }
@@ -122,7 +132,7 @@ void parse_cmd_line( int argc, char *argv[], struct opts *o )
 {
     // Collect the command line arguments
     int c;
-    while ((c = getopt( argc, argv, "a:hLo:P:t:z:")) != -1)
+    while ((c = getopt( argc, argv, "a:hLo:P:r:t:z:")) != -1)
     {
         switch (c)
         {
@@ -141,6 +151,9 @@ void parse_cmd_line( int argc, char *argv[], struct opts *o )
                 break;
             case 'P':
                 o->P_sec = atof(optarg);
+                break;
+            case 'r':
+                o->rho_lim = atof(optarg);
                 break;
             case 't':
                 o->tmult = atof(optarg);